// Home Pedia/projet/epiville/src/app/models/data-per-region.model.ts
export interface DataPerRegionResponse {
        region: string,
        securite: number | null,
        culture: number | null,
        animation: number | null,
        environnement: number | null,
        vie_pratique: number | null,
        nb_habitant: any
        superficie: any
        densite: any
        population_active: any
        taux_chomage: any
        revenu_moyen: any
        prix_m2: any
        age_population: any
}